# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: World of Shadows — Backend API
  version: 1.0.0
  description: 'Machine-readable inventory of Flask `/api/v1` routes. Human-readable request/response examples: `docs/api/REFERENCE.md`.
    Tag taxonomy: `docs/api/openapi-taxonomy.md`.'
servers:
- url: /
  description: Same origin as backend (set base URL in your client).
...
paths:
...
  /api/v1/admin/world-engine/runs:
    get:
      tags:
      - WorldEngineConsole
      summary: GET /admin/world-engine/runs
      operationId: get_api_v1_admin_world_engine_runs
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
  /api/v1/admin/world-engine/runs/{run_id}:
    get:
      tags:
      - WorldEngineConsole
      summary: GET /admin/world-engine/runs/{run_id}
      operationId: get_api_v1_admin_world_engine_runs_run_id
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
  /api/v1/admin/world-engine/runs/{run_id}/terminate:
    post:
      tags:
      - WorldEngineConsole
      summary: POST /admin/world-engine/runs/{run_id}/terminate
      operationId: post_api_v1_admin_world_engine_runs_run_id_terminate
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
  /api/v1/admin/world-engine/runs/{run_id}/transcript:
    get:
      tags:
      - WorldEngineConsole
      summary: GET /admin/world-engine/runs/{run_id}/transcript
      operationId: get_api_v1_admin_world_engine_runs_run_id_transcript
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
  /api/v1/admin/world-engine/story/sessions:
    get:
      tags:
      - WorldEngineConsole
      summary: GET /admin/world-engine/story/sessions
      operationId: get_api_v1_admin_world_engine_story_sessions
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
    post:
      tags:
      - WorldEngineConsole
      summary: POST /admin/world-engine/story/sessions
      operationId: post_api_v1_admin_world_engine_story_sessions
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
  /api/v1/admin/world-engine/story/sessions/{session_id}/diagnostics:
    get:
      tags:
      - WorldEngineConsole
      summary: GET /admin/world-engine/story/sessions/{session_id}/diagnostics
      operationId: get_api_v1_admin_world_engine_story_sessions_session_id_diagnostics
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
  /api/v1/admin/world-engine/story/sessions/{session_id}/state:
    get:
      tags:
      - WorldEngineConsole
      summary: GET /admin/world-engine/story/sessions/{session_id}/state
      operationId: get_api_v1_admin_world_engine_story_sessions_session_id_state
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
  /api/v1/admin/world-engine/story/sessions/{session_id}/turns:
    post:
      tags:
      - WorldEngineConsole
      summary: POST /admin/world-engine/story/sessions/{session_id}/turns
      operationId: post_api_v1_admin_world_engine_story_sessions_session_id_turns
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
  /api/v1/admin/world-engine/templates:
    get:
      tags:
      - WorldEngineConsole
      summary: GET /admin/world-engine/templates
      operationId: get_api_v1_admin_world_engine_templates
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
...
  /api/v1/writers-room/reviews:
    post:
      tags:
      - WritersRoom
      summary: POST /writers-room/reviews
      operationId: post_api_v1_writers_room_reviews
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
  /api/v1/writers-room/reviews/{review_id}:
    get:
      tags:
      - WritersRoom
      summary: GET /writers-room/reviews/{review_id}
      operationId: get_api_v1_writers_room_reviews_review_id
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      parameters:
      - name: review_id
        in: path
        required: true
        schema:
          type: string
  /api/v1/writers-room/reviews/{review_id}/decision:
    post:
      tags:
      - WritersRoom
      summary: POST /writers-room/reviews/{review_id}/decision
      operationId: post_api_v1_writers_room_reviews_review_id_decision
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      parameters:
      - name: review_id
        in: path
        required: true
        schema:
          type: string
  /api/v1/writers-room/reviews/{review_id}/revision-submit:
    post:
      tags:
      - WritersRoom
      summary: POST /writers-room/reviews/{review_id}/revision-submit
      operationId: post_api_v1_writers_room_reviews_review_id_revision_submit
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      parameters:
      - name: review_id
        in: path
        required: true
        schema:
          type: string
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Flask-JWT-Extended access token from `POST /api/v1/auth/login`.
security:
- BearerAuth: []